refactor(session): extract resource loader prompts
